+++ src/Web/Slack/Internal.hs
@@ -2,10 +2,11 @@
 module Web.Slack.Internal where
 
 import Data.Aeson (Value (..))
-import Network.HTTP.Client (Manager)
-import Servant.API hiding (addHeader)
 -- import Servant.Client.Core
 
+import Data.Aeson.KeyMap qualified as KM
+import Network.HTTP.Client (Manager)
+import Servant.API hiding (addHeader)
 import Servant.Client (BaseUrl (..), ClientError, ClientM, Scheme (..), mkClientEnv, runClientM)
 import Servant.Client.Core (AuthClientData, AuthenticatedRequest, Request, addHeader, mkAuthenticatedRequest)
 import Web.Slack.Common qualified as Common
@@ -17,15 +18,10 @@ data SlackConfig = SlackConfig
   , slackConfigToken :: Text
   }
 
--- contains errors that can be returned by the slack API.
--- constrast with 'SlackClientError' which additionally
--- contains errors which occured during the network communication.
-data ResponseSlackError = ResponseSlackError Text
-  deriving stock (Eq, Show)
-
 -- |
 -- Internal type!
-newtype ResponseJSON a = ResponseJSON (Either ResponseSlackError a)
+newtype ResponseJSON a = ResponseJSON (Either Common.ResponseSlackError a)
+  deriving stock (Show)
 
 instance (FromJSON a) => FromJSON (ResponseJSON a) where
   parseJSON = withObject "Response" $ \o -> do
@@ -33,7 +29,10 @@ instance (FromJSON a) => FromJSON (ResponseJSON a) where
     ResponseJSON
       <$> if ok
         then Right <$> parseJSON (Object o)
-        else Left . ResponseSlackError <$> o .: "error"
+        else do
+          err <- o .: "error"
+          meta <- o .:? "response_metadata"
+          pure $ Left $ Common.ResponseSlackError {errorText = err, responseMetadata = (fromMaybe KM.empty meta)}
 
 mkSlackAuthenticateReq :: SlackConfig -> AuthenticatedRequest (AuthProtect "token")
 mkSlackAuthenticateReq = (`mkAuthenticatedRequest` authenticateReq) . slackConfigToken
@@ -59,6 +58,6 @@ run clientAction mgr = do
 
 unnestErrors :: Either ClientError (ResponseJSON a) -> Response a
 unnestErrors (Right (ResponseJSON (Right a))) = Right a
-unnestErrors (Right (ResponseJSON (Left (ResponseSlackError serv)))) =
-  Left (Common.SlackError serv)
+unnestErrors (Right (ResponseJSON (Left err))) =
+  Left (Common.SlackError err)
 unnestErrors (Left slackErr) = Left (Common.ServantError slackErr)